+++ quotesimulator/quote_simulator_test.go
@@ -32,6 +32,7 @@ func TestSimulateQuote(t *testing.T) {
 		simulatorAddress            string
 		expectedGasAdjusted         uint64
 		expectedFeeCoin             sdk.Coin
+		expectedBaseFee             osmomath.Dec
 		expectError                 bool
 		expectedErrorMsg            string
 	}{
@@ -41,6 +42,7 @@ func TestSimulateQuote(t *testing.T) {
 			simulatorAddress:            "osmo13t8prr8hu7hkuksnfrd25vpvvnrfxr223k59ph",
 			expectedGasAdjusted:         100000,
 			expectedFeeCoin:             sdk.NewCoin("uosmo", osmomath.NewInt(10000)),
+			expectedBaseFee:             osmomath.NewDecWithPrec(5, 1),
 			expectError:                 false,
 		},
 	}
@@ -87,8 +89,12 @@ func TestSimulateQuote(t *testing.T) {
 					account *authtypes.BaseAccount,
 					chainID string,
 					msg ...sdk.Msg,
-				) (uint64, sdk.Coin, error) {
-					return tt.expectedGasAdjusted, tt.expectedFeeCoin, nil
+				) domain.QuotePriceInfo {
+					return domain.QuotePriceInfo{
+						AdjustedGasUsed: tt.expectedGasAdjusted,
+						FeeCoin:         tt.expectedFeeCoin,
+						BaseFee:         osmomath.NewDecWithPrec(5, 1),
+					}
 				},
 			}
 			txFeesClient := &mocks.TxFeesQueryClient{}
@@ -110,7 +116,7 @@ func TestSimulateQuote(t *testing.T) {
 			)
 
 			// System under test
-			gasAdjusted, feeCoin, err := simulator.SimulateQuote(
+			priceInfo := simulator.SimulateQuote(
 				context.Background(),
 				mockQuote,
 				tt.slippageToleranceMultiplier,
@@ -119,12 +125,13 @@ func TestSimulateQuote(t *testing.T) {
 
 			// Assert results
 			if tt.expectError {
-				assert.Error(t, err)
-				assert.Contains(t, err.Error(), tt.expectedErrorMsg)
+				assert.NotEmpty(t, priceInfo.Err)
+				assert.Contains(t, priceInfo.Err, tt.expectedErrorMsg)
 			} else {
-				assert.NoError(t, err)
-				assert.Equal(t, tt.expectedGasAdjusted, gasAdjusted)
-				assert.Equal(t, tt.expectedFeeCoin, feeCoin)
+				assert.Empty(t, priceInfo.Err)
+				assert.Equal(t, tt.expectedGasAdjusted, priceInfo.AdjustedGasUsed)
+				assert.Equal(t, tt.expectedFeeCoin, priceInfo.FeeCoin)
+				assert.Equal(t, tt.expectedBaseFee, priceInfo.BaseFee)
 			}
 		})
 	}